+++ include/animationPlayer.h
@@ -11,7 +11,7 @@ PlAnimation panim_createAnimation(void);
  * @param velocityY Velocity Y applied to player body before
  */
 void panim_update(PlAnimation plAnim, float velocityX, float velocityY);
-void panim_draw(PlAnimation plAnim, int posX, int posY);
+void panim_draw(PlAnimation plAnim, float posX, float posY);
 void panim_free(PlAnimation plAnim);
 
 #endif
